pyproject.toml
[tool.poetry]
name = "data-engineering-programming"
version = "0.1.0"
description = ""
authors = ["Giancarlo Lester <gatobugado.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "3.4.0"
pyarrow = "12.0.0"
grpcio = "1.48.1"
grpcio-status = "1.48.1"
protobuf = "3.19.5"
pandas = "^2.0.1"
minio = "^7.1.15"
python-decouple = "^3.8"
requests = "^2.31.0"
botocore = "^1.29.146"
boto3 = "^1.26.146"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.269"
mypy = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ['.venv', '.ruff_cache', '.mypy_cache']
ignore = []
line-length = 90
select = ['E', 'F', 'W']

[tool.black]
line-length = 90
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
    | .venv
    | .vscode
    | .ruff_cache
)
'''

[tool.mypy]
ignore_missing_imports=true